SIM: CPU trace for C64 and 1541
Added the two options
 --cpu-c64-trace TEXT        Instruction trace of the C64 6510 CPU to file
 --cpu-c1541-trace TEXT      Instruction trace of the C1541 6502 CPU to file
to get a disassembly trace of instruction stream executed by respective CPU.

Note that in order to get synthesis working the port assignment
'o_DEBUG_REGS=self.o_debug_regs' is commented out in the two wrappers since the
port is named differently depending on if we use T65 or its Verilog
translation. I.e. uncomment it for better traces in simulation.